test: split adoption tests into their own file
Keeps CaskHubTests.swift under the SwiftLint 400-line file limit.
